% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoutil.R
\name{mapLocality}
\alias{mapLocality}
\title{Leaflet-mapping of geo_df}
\usage{
mapLocality(geo_df)
}
\arguments{
\item{geo_df}{}
}
\value{
the leaflet map, for easy insertion into ggvis, if desired
}
\description{
Leaflet-mapping of geo_df
}
\examples{
mapLocality(df)
}

