% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoutil.R
\name{pointsInBox}
\alias{pointsInBox}
\title{Returns all observation inside the bounding box defined by lon/lat}
\usage{
pointsInBox(geo_df, minlon, maxlon, minlat, maxlat)
}
\arguments{
\item{geo_df}{}

\item{minlon}{}

\item{maxlon}{}

\item{minlat}{}

\item{maxlat}{}
}
\value{
data frame containing all obs inside the bounding box
}
\description{
Returns all observation inside the bounding box defined by lon/lat
}
\examples{
pointsInBox(df, 1,2,-1,1)
}

