% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoutil.R
\name{pointsInLocality}
\alias{pointsInLocality}
\title{Finds the observations in geo_df based on candidates_df}
\usage{
pointsInLocality(geo_df, candidates_df, stdDevs = 3)
}
\arguments{
\item{geo_df}{}

\item{candidates_df}{}

\item{stdDevs}{}
}
\value{
a data frame consisting of all points in the geo_df, based on a bounding box formed from the
average lon/lat of the points in candidates_df, +/- the number of stdDevs
}
\description{
Finds the observations in geo_df based on candidates_df
}
\examples{
pointsInLocality(data, data, 2)
}

