% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoutil.R
\name{visualizeLocality}
\alias{visualizeLocality}
\title{Wrapper around a ggmap-based static map. Uses google maps.}
\usage{
visualizeLocality(geo_df, zoom = 10, color = "bw")
}
\arguments{
\item{geo_df}{}

\item{zoom}{}

\item{color}{}
}
\value{
the output of ggmap
}
\description{
Wrapper around a ggmap-based static map. Uses google maps.
}
\examples{
visualizeLocality(df)
}

